+++ pkg/virt-api/webhooks/mutating-webhook/BUILD.bazel
@@ -1,53 +1,16 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = [
-        "mutating-webhook.go",
-        "namespace-limits.go",
-        "preset.go",
-    ],
+    srcs = ["mutating-webhook.go"],
     importpath = "kubevirt.io/kubevirt/pkg/virt-api/webhooks/mutating-webhook",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/api/v1:go_default_library",
         "//pkg/log:go_default_library",
         "//pkg/virt-api/webhooks:go_default_library",
+        "//pkg/virt-api/webhooks/mutating-webhook/mutators:go_default_library",
         "//pkg/virt-config:go_default_library",
         "//vendor/k8s.io/api/admission/v1beta1:go_default_library",
-        "//vendor/k8s.io/api/core/v1:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/api/resource:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/labels:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/util/errors:go_default_library",
-        "//vendor/k8s.io/client-go/tools/cache:go_default_library",
-    ],
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = [
-        "mutating-webhook_suite_test.go",
-        "mutating-webhook_test.go",
-        "namespace-limits_test.go",
-        "preset_test.go",
-    ],
-    embed = [":go_default_library"],
-    deps = [
-        "//pkg/api/v1:go_default_library",
-        "//pkg/log:go_default_library",
-        "//pkg/testutils:go_default_library",
-        "//pkg/virt-api/webhooks:go_default_library",
-        "//pkg/virt-config:go_default_library",
-        "//vendor/github.com/onsi/ginkgo:go_default_library",
-        "//vendor/github.com/onsi/gomega:go_default_library",
-        "//vendor/k8s.io/api/admission/v1beta1:go_default_library",
-        "//vendor/k8s.io/api/core/v1:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/api/resource:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/types:go_default_library",
-        "//vendor/k8s.io/client-go/tools/cache:go_default_library",
     ],
 )
