+++ pkg/virt-api/webhooks/mutating-webhook/mutating-webhook.go
@@ -21,29 +21,19 @@ package mutating_webhook
 
 import (
 	"encoding/json"
-	"fmt"
 	"net/http"
 
 	"k8s.io/api/admission/v1beta1"
-	k8sv1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 
-	v1 "kubevirt.io/kubevirt/pkg/api/v1"
 	"kubevirt.io/kubevirt/pkg/log"
 	"kubevirt.io/kubevirt/pkg/virt-api/webhooks"
+	"kubevirt.io/kubevirt/pkg/virt-api/webhooks/mutating-webhook/mutators"
 	virtconfig "kubevirt.io/kubevirt/pkg/virt-config"
 )
 
-type patchOperation struct {
-	Op    string      `json:"op"`
-	Path  string      `json:"path"`
-	Value interface{} `json:"value,omitempty"`
-}
-
 type mutator interface {
-	mutate(*v1beta1.AdmissionReview) *v1beta1.AdmissionResponse
+	Mutate(*v1beta1.AdmissionReview) *v1beta1.AdmissionResponse
 }
 
 func serve(resp http.ResponseWriter, req *http.Request, m mutator) {
@@ -55,7 +45,7 @@ func serve(resp http.ResponseWriter, req *http.Request, m mutator) {
 		return
 	}
 
-	reviewResponse := m.mutate(review)
+	reviewResponse := m.Mutate(review)
 	if reviewResponse != nil {
 		response.Response = reviewResponse
 		response.Response.UID = review.Request.UID
@@ -78,178 +68,10 @@ func serve(resp http.ResponseWriter, req *http.Request, m mutator) {
 	resp.WriteHeader(http.StatusOK)
 }
 
-type VMIsMutator struct {
-	clusterConfig *virtconfig.ClusterConfig
-}
-
-func (mutator *VMIsMutator) mutate(ar *v1beta1.AdmissionReview) *v1beta1.AdmissionResponse {
-	if ar.Request.Resource != webhooks.VirtualMachineInstanceGroupVersionResource {
-		err := fmt.Errorf("expect resource to be '%s'", webhooks.VirtualMachineInstanceGroupVersionResource.Resource)
-		return webhooks.ToAdmissionResponseError(err)
-	}
-
-	if resp := webhooks.ValidateSchema(v1.VirtualMachineInstanceGroupVersionKind, ar.Request.Object.Raw); resp != nil {
-		return resp
-	}
-
-	raw := ar.Request.Object.Raw
-	vmi := v1.VirtualMachineInstance{}
-
-	err := json.Unmarshal(raw, &vmi)
-	if err != nil {
-		return webhooks.ToAdmissionResponseError(err)
-	}
-
-	informers := webhooks.GetInformers()
-	config := mutator.clusterConfig
-
-	// Apply presets
-	err = applyPresets(&vmi, informers.VMIPresetInformer)
-	if err != nil {
-		return &v1beta1.AdmissionResponse{
-			Result: &metav1.Status{
-				Message: err.Error(),
-				Code:    http.StatusUnprocessableEntity,
-			},
-		}
-	}
-
-	// Apply namespace limits
-	applyNamespaceLimitRangeValues(&vmi, informers.NamespaceLimitsInformer)
-
-	// Set VMI defaults
-	log.Log.Object(&vmi).V(4).Info("Apply defaults")
-	setDefaultCPUModel(&vmi, config.GetCPUModel())
-	setDefaultMachineType(&vmi, config.GetMachineType())
-	setDefaultResourceRequests(&vmi, config.GetMemoryRequest(), config.GetCPURequest())
-	v1.SetObjectDefaults_VirtualMachineInstance(&vmi)
-
-	// Add foreground finalizer
-	vmi.Finalizers = append(vmi.Finalizers, v1.VirtualMachineInstanceFinalizer)
-
-	var patch []patchOperation
-	var value interface{}
-	value = vmi.Spec
-	patch = append(patch, patchOperation{
-		Op:    "replace",
-		Path:  "/spec",
-		Value: value,
-	})
-
-	value = vmi.ObjectMeta
-	patch = append(patch, patchOperation{
-		Op:    "replace",
-		Path:  "/metadata",
-		Value: value,
-	})
-
-	patchBytes, err := json.Marshal(patch)
-	if err != nil {
-		return webhooks.ToAdmissionResponseError(err)
-	}
-
-	jsonPatchType := v1beta1.PatchTypeJSONPatch
-	return &v1beta1.AdmissionResponse{
-		Allowed:   true,
-		Patch:     patchBytes,
-		PatchType: &jsonPatchType,
-	}
-
-}
-
-type MigrationCreateMutator struct {
-}
-
-func (mutator *MigrationCreateMutator) mutate(ar *v1beta1.AdmissionReview) *v1beta1.AdmissionResponse {
-	if ar.Request.Resource != webhooks.MigrationGroupVersionResource {
-		err := fmt.Errorf("expect resource to be '%s'", webhooks.MigrationGroupVersionResource.Resource)
-		return webhooks.ToAdmissionResponseError(err)
-	}
-
-	if resp := webhooks.ValidateSchema(v1.VirtualMachineInstanceMigrationGroupVersionKind, ar.Request.Object.Raw); resp != nil {
-		return resp
-	}
-
-	raw := ar.Request.Object.Raw
-	migration := v1.VirtualMachineInstanceMigration{}
-
-	err := json.Unmarshal(raw, &migration)
-	if err != nil {
-		return webhooks.ToAdmissionResponseError(err)
-	}
-
-	// Add a finalizer
-	migration.Finalizers = append(migration.Finalizers, v1.VirtualMachineInstanceMigrationFinalizer)
-	var patch []patchOperation
-	var value interface{}
-
-	value = migration.Spec
-	patch = append(patch, patchOperation{
-		Op:    "replace",
-		Path:  "/spec",
-		Value: value,
-	})
-
-	value = migration.ObjectMeta
-	patch = append(patch, patchOperation{
-		Op:    "replace",
-		Path:  "/metadata",
-		Value: value,
-	})
-
-	patchBytes, err := json.Marshal(patch)
-	if err != nil {
-		return webhooks.ToAdmissionResponseError(err)
-	}
-
-	jsonPatchType := v1beta1.PatchTypeJSONPatch
-	return &v1beta1.AdmissionResponse{
-		Allowed:   true,
-		Patch:     patchBytes,
-		PatchType: &jsonPatchType,
-	}
-}
-
-func setDefaultCPUModel(vmi *v1.VirtualMachineInstance, defaultCPUModel string) {
-	//if vmi doesn't have cpu topology or cpu model set
-	if vmi.Spec.Domain.CPU == nil || vmi.Spec.Domain.CPU.Model == "" {
-		if defaultCPUModel != "" {
-			// create cpu topology struct
-			if vmi.Spec.Domain.CPU == nil {
-				vmi.Spec.Domain.CPU = &v1.CPU{}
-			}
-			//set is as vmi cpu model
-			vmi.Spec.Domain.CPU.Model = defaultCPUModel
-		}
-	}
-}
-
-func setDefaultMachineType(vmi *v1.VirtualMachineInstance, defaultMachineType string) {
-	if vmi.Spec.Domain.Machine.Type == "" {
-		vmi.Spec.Domain.Machine.Type = defaultMachineType
-	}
-}
-
-func setDefaultResourceRequests(vmi *v1.VirtualMachineInstance, defaultMemoryRequest resource.Quantity, defaultCPURequest resource.Quantity) {
-	if _, exists := vmi.Spec.Domain.Resources.Requests[k8sv1.ResourceMemory]; !exists {
-		if vmi.Spec.Domain.Resources.Requests == nil {
-			vmi.Spec.Domain.Resources.Requests = k8sv1.ResourceList{}
-		}
-		vmi.Spec.Domain.Resources.Requests[k8sv1.ResourceMemory] = defaultMemoryRequest
-	}
-
-	if _, exists := vmi.Spec.Domain.Resources.Requests[k8sv1.ResourceCPU]; !exists {
-		if vmi.Spec.Domain.CPU != nil && vmi.Spec.Domain.CPU.DedicatedCPUPlacement {
-			return
-		}
-		vmi.Spec.Domain.Resources.Requests[k8sv1.ResourceCPU] = defaultCPURequest
-	}
-}
-
 func ServeVMIs(resp http.ResponseWriter, req *http.Request, clusterConfig *virtconfig.ClusterConfig) {
-	serve(resp, req, &VMIsMutator{clusterConfig: clusterConfig})
+	serve(resp, req, &mutators.VMIsMutator{ClusterConfig: clusterConfig})
 }
 
 func ServeMigrationCreate(resp http.ResponseWriter, req *http.Request) {
-	serve(resp, req, &MigrationCreateMutator{})
+	serve(resp, req, &mutators.MigrationCreateMutator{})
 }
