+++ pkg/virt-api/webhooks/mutating-webhook/mutators/mutators_suite_test.go
@@ -1,4 +1,4 @@
-package mutating_webhook
+package mutators
 
 import (
 	"testing"
@@ -12,5 +12,5 @@ import (
 func TestMutatingWebhook(t *testing.T) {
 	log.Log.SetIOWriter(GinkgoWriter)
 	RegisterFailHandler(Fail)
-	RunSpecs(t, "MutatingWebhook Suite")
+	RunSpecs(t, "Mutators Suite")
 }
