+++ pkg/virt-api/webhooks/mutating-webhook/mutators/namespace-limits_test.go
@@ -17,7 +17,7 @@
  *
  */
 
-package mutating_webhook
+package mutators
 
 import (
 	. "github.com/onsi/ginkgo"
