+++ pkg/virt-api/webhooks/mutating-webhook/mutators/preset_test.go
@@ -17,7 +17,7 @@
  *
  */
 
-package mutating_webhook
+package mutators
 
 import (
 	"fmt"
